Trait fastfield_codecs::MonotonicallyMappableToU128

pub trait MonotonicallyMappableToU128: 'static + PartialOrd + Copy + Send + Sync + Debug {
    fn to_u128(self) -> u128;
    fn from_u128(val: u128) -> Self;
}

Montonic maps a value to u128 value space Monotonic mapping enables PartialOrd on u128 space without conversion to original space.

Required Methods

fn to_u128(self) -> u128

Converts a value to u128.

Internally all fast field values are encoded as u64.

fn from_u128(val: u128) -> Self

Converts a value from u128

Internally all fast field values are encoded as u64. Note: To be used for converting encoded Term, Posting values.