Crate fastfield_codecs

fastfield_codecs

Columnar storage of data for tantivy Column.

Encode data in different codecs.

Monotonically map values to u64/u128

Structs

IterColumn

Wraps an iterator into a Column.

NormalizedHeader

The normalized header gives some parameters after applying the following normalization of the vector: val -> (val - min_value) / gcd

VecColumn

VecColumn provides Column over a slice.

Enums

FastFieldCodecType

Available codecs to use to encode the u64 (via MonotonicallyMappableToU64) converted data.

U128FastFieldCodecType

Available codecs to use to encode the u128 (via MonotonicallyMappableToU128) converted data.

Constants

ALL_CODEC_TYPES

The list of all available codecs for u64 convertible data.

Traits

Column

Column provides columnar access on a field.

MonotonicallyMappableToU64

Monotonic maps a value to u64 value space. Monotonic mapping enables PartialOrd on u64 space without conversion to original space.

MonotonicallyMappableToU128

Montonic maps a value to u128 value space Monotonic mapping enables PartialOrd on u128 space without conversion to original space.

StrictlyMonotonicFn

Values need to be strictly monotonic mapped to a Internal value (u64 or u128) that can be used in fast field codecs.

Functions

estimate

Return estimated compression for given codec in the value range [0.0..1.0], where 1.0 means no compression.

monotonic_map_column

Creates a view of a column transformed by a strictly monotonic mapping. See StrictlyMonotonicFn.

open

Returns the correct codec reader wrapped in the Arc for the data.

open_u128

Returns the correct codec reader wrapped in the Arc for the data.

serialize

Serializes the column with the codec with the best estimate on the data.

serialize_and_load

Helper function to serialize a column (autodetect from all codecs) and then open it

serialize_u128

Serializes u128 values with the compact space codec.