Trait fastfield_codecs::MonotonicallyMappableToU64

pub trait MonotonicallyMappableToU64: 'static + PartialOrd + Copy + Send + Sync + Debug {
    // Required methods
    fn to_u64(self) -> u64;
    fn from_u64(val: u64) -> Self;
}

Monotonic maps a value to u64 value space. Monotonic mapping enables PartialOrd on u64 space without conversion to original space.

Required Methods

fn to_u64(self) -> u64

Converts a value to u64.

Internally all fast field values are encoded as u64.

fn from_u64(val: u64) -> Self

Converts a value from u64

Internally all fast field values are encoded as u64. Note: To be used for converting encoded Term, Posting values.